# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FieldServicer Support Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List outbound webhooks
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Create an outbound webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookBody"
      responses:
        "201":
          description: Created webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /webhooks/{id}:
    patch:
      operationId: updateWebhook
      tags: [webhooks]
      summary: Update a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookBody"
      responses:
        "200":
          description: Updated webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        secret:
          type: string
          nullable: true
        active:
          type: boolean
        lastTriggeredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, url, events, active, createdAt]
    CreateWebhookBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        secret:
          type: string
        active:
          type: boolean
      required: [name, url]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]