# lib/api-zod/src/generated/types/createWebhookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FieldServicer Support Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWebhookBody {
  name: string;
  url: string;
  events?: string[];
  secret?: string;
  active?: boolean;
}